+++ src/common/OperationsBase.ts
@@ -15,7 +15,7 @@ const MAX_ATTEMPTS = 3;
 export class OperationsBase {
   protected readonly fetch = isomorphicFetch;
 
-  constructor(protected readonly basePath?: string) { }
+  constructor(protected readonly basePath: string) { }
 
   /**
    * Creates a request body and headers