+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 Thursday, 19 September, 2024
 ### Minor
 - Added `NamedGroupsClient`.
+- `OperationsBase` now requires a base path.
 
 ## 0.13.0
 Tuesday, 10 September, 2024